requirements.txt
# Telegram Bot
python-telegram-bot[job-queue]==21.7

# Google Gemini AI
google-generativeai>=0.8.0

# Poe API (OpenAI-compatible)
openai>=1.0.0

# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Environment
python-dotenv>=1.0.0
